+++ src/components/UserCard.tsx
@@ -13,7 +13,15 @@ import { useUserStore } from "@/store/userStore";
 //   skills: string[];
 // }
 
-export const UserCard = ({ users }: { users: UserData[] }) => {
+interface UserCardProps {
+  users: UserData[];
+  isDisableSuggest?: boolean;
+}
+
+export const UserCard = ({
+  users,
+  isDisableSuggest = false,
+}: UserCardProps) => {
   const navigate = useNavigate();
   const { onOpenModal } = useModalStore();
   const { userId } = useUserStore();
@@ -47,8 +55,19 @@ export const UserCard = ({ users }: { users: UserData[] }) => {
               />
             ) : (
               <span className="h-12 w-12 flex items-center justify-center bg-gray-200 rounded-full mb-3">
-                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="h-8 w-8 text-gray-400">
-                  <path strokeLinecap="round" strokeLinejoin="round" d="M15.75 6a3.75 3.75 0 1 1-7.5 0 3.75 3.75 0 0 1 7.5 0ZM4.501 20.118a7.5 7.5 0 0 1 14.998 0A17.933 17.933 0 0 1 12 21.75c-2.676 0-5.216-.584-7.499-1.632Z" />
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  strokeWidth="1.5"
+                  stroke="currentColor"
+                  className="h-8 w-8 text-gray-400"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    d="M15.75 6a3.75 3.75 0 1 1-7.5 0 3.75 3.75 0 0 1 7.5 0ZM4.501 20.118a7.5 7.5 0 0 1 14.998 0A17.933 17.933 0 0 1 12 21.75c-2.676 0-5.216-.584-7.499-1.632Z"
+                  />
                 </svg>
               </span>
             )}
@@ -87,6 +106,7 @@ export const UserCard = ({ users }: { users: UserData[] }) => {
               <Button
                 onClick={() => handleOpenSuggest(user.userId)}
                 color={"blue"}
+                disabled={isDisableSuggest || userId === user.userId}
               >
                 제안하기
               </Button>
